model/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pandas==2.2.3
python-multipart==0.0.20
autogluon.tabular
